% Auto-generated: do not edit by hand
\name{''FeetSensors}

\alias{''FeetSensors}

\title{FeetSensors component}

\description{
Custom component of 2 feet with 3 sensors on each foot.
}

\usage{
''FeetSensors(id=NULL, L0=NULL, L1=NULL, L2=NULL, R0=NULL, R1=NULL,
R2=NULL, anomaly_L0=NULL, anomaly_L1=NULL, anomaly_L2=NULL,
anomaly_R0=NULL, anomaly_R1=NULL, anomaly_R2=NULL)
}

\arguments{
\item{id}{Character. Component ID.}

\item{L0}{Numeric. Value of the L0 sensor.}

\item{L1}{Numeric. Value of the L0 sensor.}

\item{L2}{Numeric. Value of the L0 sensor.}

\item{R0}{Numeric. Value of the L0 sensor.}

\item{R1}{Numeric. Value of the L0 sensor.}

\item{R2}{Numeric. Value of the L0 sensor.}

\item{anomaly_L0}{Logical. Boolean value informing if there is an anomaly on the L0 sensor.}

\item{anomaly_L1}{Logical. Boolean value informing if there is an anomaly on the L0 sensor.}

\item{anomaly_L2}{Logical. Boolean value informing if there is an anomaly on the L0 sensor.}

\item{anomaly_R0}{Logical. Boolean value informing if there is an anomaly on the L0 sensor.}

\item{anomaly_R1}{Logical. Boolean value informing if there is an anomaly on the L0 sensor.}

\item{anomaly_R2}{Logical. Boolean value informing if there is an anomaly on the L0 sensor.}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

