% Auto-generated: do not edit by hand
\name{''FeetSensors}

\alias{''FeetSensors}

\title{FeetSensors component}

\description{

}

\usage{
''FeetSensors(id=NULL, L0=NULL, L1=NULL, L2=NULL, R0=NULL, R1=NULL,
R2=NULL, anomaly_L0=NULL, anomaly_L1=NULL, anomaly_L2=NULL,
anomaly_R0=NULL, anomaly_R1=NULL, anomaly_R2=NULL)
}

\arguments{
\item{id}{Character. }

\item{L0}{Numeric. }

\item{L1}{Numeric. }

\item{L2}{Numeric. }

\item{R0}{Numeric. }

\item{R1}{Numeric. }

\item{R2}{Numeric. }

\item{anomaly_L0}{Logical. }

\item{anomaly_L1}{Logical. }

\item{anomaly_L2}{Logical. }

\item{anomaly_R0}{Logical. }

\item{anomaly_R1}{Logical. }

\item{anomaly_R2}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

